Move dynamo functions and address review comments.
